tidyrules

tidyrules package provides a framework to work with decision rules. Rules can be extracted from supported models using tidy, augmented using validation data by augment, manipulated using standard dataframe operations, (modified) rulelists can be used to predict on unseen (test) data. Utilities include: Create a rulelist manually (as_rulelist), Export a rulelist to SQL (to_sql_case) and so on. The package offers two classes; rulelist and ruleset based on dataframe.

See also

Author

Maintainer: Srikanth Komala Sheshachala sri.teach@gmail.com

Authors:

Amith Kumar Ullur Raghavendra amith54@gmail.com